Вынести классы Node и TasksLinkedList в отдельные файлы. Исправить ошибку CheckStyle.
