+++ src/service/Node.java
@@ -1,0 +1,11 @@
+package service;
+
+class Node<T> {
+    public T data;
+    public Node<T> next;
+    public Node<T> prev;
+
+    public Node(T data) {
+        this.data = data;
+    }
+}
